refactor: move portforward to AbstractOperatorExtension
